Refactor : 위치 변경
